* [FILE] renamed the stbImage folder to STB * [GRAPHICS] added basic functionality to the font asset * [LIBRARY] added stb_truetype and stb_rect_pack for the fonts * [GRAPHICS] made the font manager empty for later usage (maybe)

+++ Engine/ChiliMath.h
@@ -59,7 +59,7 @@ inline float Rad2Def(float rad)
 }
 inline float Log2(float x)
 {
-	return log(x) * INV_LOG2;
+	return logf(x) * INV_LOG2;
 }
 inline int Log2int(float x)
 {
